RR-584 - Refactor Work & Skills cypress tests into separate files per "card"
